backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pandas==2.2.2
scipy==1.13.1
scikit-learn==1.5.1
requests==2.32.3
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
reportlab==4.2.2